# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subscription": {
      "post": {
        "description": "Creates (or reactivates) the subscription for the authenticated user and starts payment for the selected plan. A trial is granted immediately and expires after TRIAL_DAYS. A paid plan is applied only once payment settles: the response is 200 when it has, and 202 with a checkoutUrl to visit when it has not. Limits are enforced from that point on: max_organizations, max_users, max_projects.",
        "operationId": "SubscriptionController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment settled; the plan is in force",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionCreatedResponseDto"
                }
              }
            }
          },
          "202": {
            "description": "Payment started but not settled; complete the checkout before the plan applies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchasePendingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting state: the resource already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Subscribe to a plan",
        "tags": [
          "subscription"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      },
      "get": {
        "description": "Returns the subscription of the authenticated user with the attached plan and current usage against every plan limit.",
        "operationId": "SubscriptionController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current subscription and usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionWithUsageDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current subscription",
        "tags": [
          "subscription"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      },
      "delete": {
        "description": "Cancels the subscription. Data is preserved but capacity-consuming operations are refused until the subscription is reactivated by subscribing again.",
        "operationId": "SubscriptionController_cancel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel subscription",
        "tags": [
          "subscription"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      }
    },
    "/subscription/usage": {
      "get": {
        "description": "Capacity consumed versus granted for organizations, users and projects.",
        "operationId": "SubscriptionController_usage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usage per limited resource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanUsageDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get plan usage",
        "tags": [
          "subscription"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      }
    },
    "/subscription/plan": {
      "patch": {
        "description": "Starts payment for another plan. The current plan stays in force until the payment settles, so a plan is never granted without money behind it. Downgrades are rejected up front, while the current usage exceeds the target plan limits. The response is 200 when payment settled and 202 with a checkoutUrl when the customer has to complete a checkout first.",
        "operationId": "SubscriptionController_changePlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment settled; the new plan is in force",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionWithUsageDto"
                }
              }
            }
          },
          "202": {
            "description": "Payment started but not settled; complete the checkout before the plan applies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchasePendingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change plan",
        "tags": [
          "subscription"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      }
    },
...
  },
  "info": {
    "title": "OPS API",
...
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Validation failed"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "name should not be empty",
                  "email must be an email"
                ]
              }
            ],
            "description": "Human readable summary: one string, or one string per invalid field"
          },
          "error": {
            "type": "string",
            "enum": [
              "Bad Request",
              "Unauthorized",
              "Forbidden",
              "Not Found",
              "Conflict",
              "Too Many Requests"
            ],
            "example": "Bad Request",
            "description": "Status name. Absent on the 401 of a rejected or expired JWT and on every 429, because those exceptions are thrown without a message."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b6f0e2a1-9f2a-4a6f-8f1e-2c9a5b7d1e33"
          },
          "name": {
            "type": "string",
            "example": "Growth"
          },
          "description": {
            "type": "object",
            "example": "Multiple workspaces for growing delivery teams."
          },
          "maxUsers": {
            "type": "number",
            "example": 25,
            "description": "Maximum users permitted"
          },
          "maxProjects": {
            "type": "number",
            "example": 25,
            "description": "Maximum projects permitted"
          },
          "maxStorageGb": {
            "type": "number",
            "example": 50,
            "description": "Maximum storage in GB"
          },
          "maxOrganizations": {
            "type": "number",
            "example": 3,
            "description": "Maximum organizations permitted"
          },
          "price": {
            "type": "string",
            "example": "79.00",
            "description": "Subscription price (DECIMAL(10,2))"
          }
        },
        "required": [
          "id",
          "name",
          "maxUsers",
          "maxProjects",
          "maxStorageGb",
          "maxOrganizations",
          "price"
        ]
      },
      "SelectPlanDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "example": "b6f0e2a1-9f2a-4a6f-8f1e-2c9a5b7d1e33",
            "description": "Identifier of the selected plan (see GET /plans)"
          },
          "status": {
            "type": "string",
            "enum": [
              "trial",
              "active"
            ],
            "default": "active",
            "description": "Subscription status. \"trial\" starts a trial subscription; \"active\" activates immediately. A client cannot subscribe straight into a cancelled or suspended state."
          }
        },
        "required": [
          "planId"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4c1f5c66-2f9e-4f5c-8b2a-9c0d1e2f3a4b"
          },
          "userId": {
            "type": "string",
            "example": "9a2b7c1d-4e5f-4a6b-8c9d-0e1f2a3b4c5d"
          },
          "planId": {
            "type": "object",
            "example": "b6f0e2a1-9f2a-4a6f-8f1e-2c9a5b7d1e33",
            "description": "Plan in force. Null before the first payment settles and while a plan change is still being paid for, in which case the previous plan is also null and pendingPlan names the one being bought."
          },
          "plan": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanResponseDto"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "trial",
              "pending_payment",
              "active",
              "cancelled",
              "suspended"
            ],
            "example": "active"
          },
          "trialEndsAt": {
            "type": "object",
            "example": "2026-10-16T10:00:00.000Z",
            "description": "When a trial stops granting capacity. Null unless the subscription was started as a trial."
          },
          "isTrialExpired": {
            "type": "boolean",
            "example": false,
            "description": "True once the trial has passed. A lapsed trial keeps its data and stops accepting organizations, members and projects until a plan is chosen."
          },
          "pendingPlan": {
            "nullable": true,
            "description": "Plan currently being paid for, or null when no purchase is in flight.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanResponseDto"
              }
            ]
          },
          "billingSubscriptionRef": {
            "type": "object",
            "example": "sub_1234abcd",
            "nullable": true,
            "description": "Payment provider handle for the recurring agreement"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-30T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-30T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "status",
          "isTrialExpired",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlanUsageDto": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "enum": [
              "organizations",
              "users",
              "projects"
            ],
            "example": "organizations",
            "description": "Limited resource"
          },
          "used": {
            "type": "number",
            "example": 1,
            "description": "Capacity currently consumed"
          },
          "limit": {
            "type": "number",
            "example": 3,
            "description": "Limit granted by the plan"
          },
          "remaining": {
            "type": "number",
            "example": 2,
            "description": "Remaining capacity"
          }
        },
        "required": [
          "resource",
          "used",
          "limit",
          "remaining"
        ]
      },
      "PurchaseStateDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "settled"
            ],
            "example": "pending"
          },
          "checkoutUrl": {
            "type": "object",
            "example": "https://checkout.example.com/session/cs_test_123",
            "nullable": true,
            "description": "Where to complete payment. Present only while the purchase is pending."
          }
        },
        "required": [
          "status"
        ]
      },
      "SubscriptionCreatedResponseDto": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionResponseDto"
          },
          "usage": {
            "description": "Current usage against plan limits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanUsageDto"
            }
          },
          "purchase": {
            "description": "Payment state of the most recent plan change. Absent when the plan was not changed.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PurchaseStateDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Subscription active"
          }
        },
        "required": [
          "subscription",
          "usage",
          "message"
        ]
      },
      "PurchasePendingResponseDto": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionResponseDto"
          },
          "usage": {
            "description": "Current usage against plan limits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanUsageDto"
            }
          },
          "purchase": {
            "description": "Payment state of the most recent plan change. Absent when the plan was not changed.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PurchaseStateDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Complete the payment to activate this plan"
          }
        },
        "required": [
          "subscription",
          "usage",
          "message"
        ]
      },
      "SubscriptionWithUsageDto": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionResponseDto"
          },
          "usage": {
            "description": "Current usage against plan limits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanUsageDto"
            }
          },
          "purchase": {
            "description": "Payment state of the most recent plan change. Absent when the plan was not changed.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PurchaseStateDto"
              }
            ]
          }
        },
        "required": [
          "subscription",
          "usage"
        ]
      },
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "example": "b6f0e2a1-9f2a-4a6f-8f1e-2c9a5b7d1e33"
          }
        },
        "required": [
          "planId"
        ]
      },
      "SubscriptionMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Subscription cancelled"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}